requirements.txt
# Core Telegram Bot dependencies
python-telegram-bot==21.0.1
httpx==0.28.1

# CryptoRank API dependencies  
aiohttp==3.11.12
requests==2.32.3

# Configuration and environment
python-dotenv==1.0.1

# AI Model dependencies (required if using MODEL_API_KEY)
openai==1.59.9
langchain-core==0.3.30

# Core Python dependencies (usually included but ensuring compatibility)
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
anyio==4.8.0
sniffio==1.3.1
h11==0.14.0
httpcore==1.0.7